Rename vhu to bsvhu
