+++ back/src/bsds/resolvers/queries/bsds.ts
@@ -17,8 +17,8 @@ import {
 } from "../../../common/elastic";
 import prisma from "../../../prisma";
 import { expandFormFromDb } from "../../../forms/form-converter";
-import { expandBsdasriFromDb } from "../../../bsdasris/dasri-converter";
-import { expandVhuFormFromDb } from "../../../vhu/converter";
+import { expandBsdasriFromDb } from "../../../dasris/dasri-converter";
+import { expandVhuFormFromDb } from "../../../bsvhu/converter";
 import { expandBsdaFromDb } from "../../../bsda/converter";
 import { getUserCompanies } from "../../../users/database";
