+++ back/src/bsffs/resolvers/queries/bsffs.ts
@@ -5,7 +5,7 @@ import { checkIsAuthenticated } from "../../../common/permissions";
 import { getUserCompanies } from "../../../users/database";
 import { safeInput } from "../../../forms/form-converter";
 import { Prisma } from ".prisma/client";
-import { getConnectionsArgs } from "../../../vhu/pagination";
+import { getConnectionsArgs } from "../../../bsvhu/pagination";
 
 const bsffs: QueryResolvers["bsffs"] = async (_, args, context) => {
   const user = checkIsAuthenticated(context);
