+++ back/src/bsvhu/errors.ts
