+++ back/src/bsvhu/resolvers/mutations/delete.ts
