+++ back/src/bsvhu/typeDefs/bsvhu.enums.graphql
