+++ back/src/bsvhu/typeDefs/bsvhu.inputs.graphql
