+++ back/src/bsvhu/validation.ts
