+++ back/src/schema.ts
@@ -4,7 +4,7 @@ import scalarResolvers from "./scalars";
 import companiesResolvers from "./companies/resolvers";
 import usersResolvers from "./users/resolvers";
 import formsResolvers from "./forms/resolvers";
-import vhuResolvers from "./vhu/resolvers";
+import vhuResolvers from "./bsvhu/resolvers";
 import bsdsResolvers from "./bsds/resolvers";
 import dasriResolvers from "./bsdasris/resolvers";
 import bsffResolvers from "./bsffs/resolvers";
@@ -18,7 +18,7 @@ const repositories = [
   "users",
   "companies",
   "forms",
-  "vhu",
+  "bsvhu",
   "bsds",
   "bsdasris",
   "bsffs",
