+++ front/src/form/bsvhu/utils/queries.ts
@@ -126,8 +126,8 @@ export const DUPLICATE_VHU_FORM = gql`
 `;
 
 export const CREATE_VHU_FORM = gql`
-  mutation CreateBsvhu($input: BsvhuInput!) {
-    createBsvhu(input: $input) {
+  mutation CreateDraftBsvhu($input: BsvhuInput!) {
+    createDraftBsvhu(input: $input) {
       ...FullBsvhu
     }
   }
